fix: improve data quality

+++ data/processed/f_768_wenhao_wo_doc.py
@@ -57,7 +57,8 @@ class TestCases(unittest.TestCase):
     def tearDown(self):
         os.remove(self.db1_path)
         os.remove(self.db2_path)
-        os.remove('nonexistent.db')
+        if os.path.exists('nonexistent.db'):
+            os.remove('nonexistent.db')
     
     def test_valid_input(self):
         df1 = f_417(self.db1_path, self.table_name1, self.column_name1)
